samples/python/agents/langchain/pyproject.toml
[project]
name = "a2a-sample-agent-langchain"
version = "0.1.0"
description = "Sample Langchain agent with A2A Protocol"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.1.8",
    "httpx>=0.28.1",
    "langchain>=0.2.0",
    "python-dotenv>=1.1.0",
    "langchain-core>=0.2.0",
    "langchain-ollama>=0.3.2",
    "a2a-samples",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.uv.sources]
a2a-samples = { workspace = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"